reorganizing modChannelView
